test for lif_synapse done
