+++ Network/STDP.py
@@ -3,7 +3,6 @@ import os
 
 sys.path.append(os.path.join(os.path.dirname(os.path.dirname(os.path.abspath(__file__))), "Neurons"))
 
-from Neuron import Neuron as Neuron_
 
 import math
 import numpy as np
