+++ Neurons/test/LIF_synapse-test.py
@@ -1,26 +1,22 @@
-from __init__ import * 
-from IZHI import IZHI 
+from __init__ import *
+from LIF_synapse import *  
 from plot import plot
 
 '''Test EULER METHOD
 
 There's unity test, it's necessary to compare the output image with the image from the folder expected-results
 
-Parameters for test: 
+Paramter for test: 
 
-    a = .1            
-    b = .2
-    c = -65
-    d = 8
-    
-    thrs = 30 
-    v = -65
-    u = -14
-
+    self.R = 5  
+    self.C = 3
+    self.uR = -40
+    self.thrs = 30
+    self.maxV = 90
+    self.v = -65
 '''
-
-
-neuron = IZHI() 
+ 
+neuron = LIF() 
 time = 500
 dt = 0.01
 steps = math.ceil(time / dt)
@@ -29,7 +25,7 @@ I = [0 if 200/dt <= i <= 300/dt  else 10 for i in range(steps)]
 # Comment the test you do not wish to do 
 
 # TEST : EULER
-plot(neuron, time, dt, I, 0)
+#plot(neuron, time, dt, I, 0)
 
 #TEST : RK4
-# plot(neuron, time, dt, I, 1)
+plot(neuron, time, dt, I, 1)
